controle de fluxos

+++ app/Http/Controllers/MarcaController.php
@@ -7,7 +7,8 @@ use Illuminate\Http\Request;
 
 class MarcaController extends Controller
 {
-    public function __construct(Marca $marca) {
+    public function __construct(Marca $marca)
+    {
         $this->marca = $marca;
     }
     /**
@@ -43,6 +44,9 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
+        if ($marca === null) {
+            return ['erro' => 'Recurso pesquisado não existe'];
+        }
         return $marca;
     }
 
@@ -57,6 +61,11 @@ class MarcaController extends Controller
     {
         //$marca->update($request->all());
         $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return ['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'];
+        }
+
         $marca->update($request->all());
         return $marca;
     }
@@ -72,6 +81,11 @@ class MarcaController extends Controller
         //$marca->delete();
 
         $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return ['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe'];
+        }
+
         $marca->delete();
         return ['msg' => 'A marca foi removida com sucesso'];
     }
